Create routes.js file
